Arreglo fallitos

+++ Item 4/Acme-Flights/src/main/java/controllers/manager/CampaignManagerController.java
@@ -15,10 +15,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
 import services.CampaignService;
+import services.ExchangeRateService;
 import services.ManagerService;
+import services.MonthlyBillService;
 import controllers.AbstractController;
 import domain.Campaign;
+import domain.ExchangeRate;
 import domain.Manager;
+import domain.MonthlyBill;
 
 @Controller
 @RequestMapping("/campaign/manager")
@@ -26,10 +30,16 @@ public class CampaignManagerController extends AbstractController {
 
 	// Services ---------------------------------------------------------------
 	@Autowired
-	private CampaignService	campaignService;
+	private CampaignService		campaignService;
 
 	@Autowired
-	private ManagerService	managerService;
+	private ManagerService		managerService;
+
+	@Autowired
+	private MonthlyBillService	monthlyBillService;
+
+	@Autowired
+	private ExchangeRateService	exchangeRateService;
 
 
 	// Constructors -----------------------------------------------------------
@@ -123,16 +133,31 @@ public class CampaignManagerController extends AbstractController {
 
 		campaign = this.campaignService.findOne(campaignId);
 
-		try {
-			this.campaignService.delete(campaign);
-			result = new ModelAndView("redirect:../manager/list.do?");
-		} catch (final Throwable oops) {
-			result = this.createEditModelAndView(campaign, "campaign.commit.error");
-		}
+		if (this.campaignService.findUnpaidMonthlyBillsByCampaignId(campaign.getId()).size() > 0) {
+			Collection<MonthlyBill> monthlyBills;
+			Manager manager;
+			Collection<ExchangeRate> exchangeRates;
+
+			manager = this.managerService.findByPrincipal();
+
+			exchangeRates = this.exchangeRateService.findAll();
+			monthlyBills = this.monthlyBillService.findByAirlineId(manager.getAirline().getId());
+
+			result = new ModelAndView("monthlyBill/list");
+			result.addObject("monthlyBills", monthlyBills);
+			result.addObject("requestURI", "monthlyBill/manager/list.do");
+			result.addObject("exchangeRates", exchangeRates);
+			result.addObject("errorUnpaid", true);
+		} else
+			try {
+				this.campaignService.delete(campaign);
+				result = new ModelAndView("redirect:list.do");
+			} catch (final Throwable oops) {
+				result = this.createEditModelAndView(campaign, "campaign.commit.error");
+			}
 
 		return result;
 	}
-
 	// Ancillary methods ------------------------------------------------------
 
 	protected ModelAndView createEditModelAndView(final Campaign campaign) {
